+++ cookbooks/chub_castle/attributes/default.rb
@@ -4,3 +4,6 @@ default[:chub_castle][:keystore_file] = "keystore.jks"
 default[:chub_castle][:truststore_file] = "truststore.jks"
 default[:chub_castle][:keystore_password] = "changeit"
 default[:chub_castle][:truststore_password] = "changeit"
+default[:chub_castle][:hazelcast][:group] = "#{node.chef_environment}-castle"
+default[:chub_castle][:hazelcast][:service_ticket_timeout] = 300
+default[:chub_castle][:hazelcast][:ticket_granting_ticket_timeout] = 7200