+++ cookbooks/chub_plaza/environments/ssodev.rb
@@ -1,19 +1,5 @@
 name "dev"
 default_attributes(
-  :java => {
-    :oracle => {
-      :accept_oracle_download_terms => true
-    },
-    :install_flavor => "oracle",
-    :jdk_version => "8",
-    :jdk => {
-      "8" => {
-        :x86_64 => {
-          :url => "http://artifactory01.nexus.commercehub.com/artifactory/java-distributions/jdk/8-b132/jdk-8-linux-x64.tar.gz"
-        }
-      }
-    }
-  },
   :sensu => {
     :rabbitmq => {
       :host => "192.168.24.132"